processing/requirements.txt
numpy
torch
scipy
open3d
libigl==2.4.1
polyscope